+++ src/main/java/com/leadlet/service/impl/TeamServiceImpl.java
@@ -1,5 +1,6 @@
 package com.leadlet.service.impl;
 
+import com.leadlet.security.SecurityUtils;
 import com.leadlet.service.TeamService;
 import com.leadlet.domain.Team;
 import com.leadlet.repository.TeamRepository;
@@ -12,13 +13,15 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
+
 
 /**
  * Service Implementation for managing Team.
  */
 @Service
 @Transactional
-public class TeamServiceImpl implements TeamService{
+public class TeamServiceImpl implements TeamService {
 
     private final Logger log = LoggerFactory.getLogger(TeamServiceImpl.class);
 
@@ -41,46 +44,75 @@ public class TeamServiceImpl implements TeamService{
     public TeamDTO save(TeamDTO teamDTO) {
         log.debug("Request to save Team : {}", teamDTO);
         Team team = teamMapper.toEntity(teamDTO);
+        team.setAppAccount(SecurityUtils.getCurrentUserAppAccount());
         team = teamRepository.save(team);
         return teamMapper.toDto(team);
     }
 
     /**
-     *  Get all the teams.
+     * Update a team.
      *
-     *  @param pageable the pagination information
-     *  @return the list of entities
+     * @param teamDTO the entity to update
+     * @return the persisted entity
+     */
+    @Override
+    public TeamDTO update(TeamDTO teamDTO) {
+        log.debug("Request to save Team : {}", teamDTO);
+        Team team = teamMapper.toEntity(teamDTO);
+
+        Team teamFromDb = teamRepository.findOneByIdAndAppAccount(team.getId(), SecurityUtils.getCurrentUserAppAccount());
+
+        if (teamFromDb != null) {
+            team.setAppAccount(SecurityUtils.getCurrentUserAppAccount());
+            team = teamRepository.save(team);
+            return teamMapper.toDto(team);
+        } else {
+            throw new EntityNotFoundException();
+        }
+    }
+
+    /**
+     * Get all the teams.
+     *
+     * @param pageable the pagination information
+     * @return the list of entities
      */
     @Override
     @Transactional(readOnly = true)
     public Page<TeamDTO> findAll(Pageable pageable) {
         log.debug("Request to get all Teams");
-        return teamRepository.findAll(pageable)
+        return teamRepository.findByAppAccount(SecurityUtils.getCurrentUserAppAccount(), pageable)
             .map(teamMapper::toDto);
     }
 
     /**
-     *  Get one team by id.
+     * Get one team by id.
      *
-     *  @param id the id of the entity
-     *  @return the entity
+     * @param id the id of the entity
+     * @return the entity
      */
     @Override
     @Transactional(readOnly = true)
     public TeamDTO findOne(Long id) {
         log.debug("Request to get Team : {}", id);
-        Team team = teamRepository.findOne(id);
+        Team team = teamRepository.findOneByIdAndAppAccount(id, SecurityUtils.getCurrentUserAppAccount());
         return teamMapper.toDto(team);
     }
 
     /**
-     *  Delete the  team by id.
+     * Delete the  team by id.
      *
-     *  @param id the id of the entity
+     * @param id the id of the entity
      */
     @Override
     public void delete(Long id) {
         log.debug("Request to delete Team : {}", id);
-        teamRepository.delete(id);
+        Team teamFromDb = teamRepository.findOneByIdAndAppAccount(id, SecurityUtils.getCurrentUserAppAccount());
+
+        if (teamFromDb != null) {
+            teamRepository.delete(id);
+        } else {
+            throw new EntityNotFoundException();
+        }
     }
 }
